rework physics

+++ Player.h
@@ -9,7 +9,7 @@
 class Player : public PhysicsObject {
 public:
     bool is_player = true;
-    Player(char sprite, int x, int y);
+    Player(char sprite, float x, float y);
     void tick() override;
     bool check_collision(Player *p) override;
     bool collide(Player *p) override;
