+++ Scene.h
@@ -26,12 +26,12 @@ public:
     Scene(const Scene &other) = delete; // copy ctor
     void operator=(const Scene &other) = delete; // copy assignment
 
-    void new_sprite(char sprite, int x = 0, int y = 0);
-    void new_collectible(char sprite, int x = 0, int y = 0);
-    void new_end(char sprite, int x = 0, int y = 0);
-    void new_physics(char sprite, int x = 0, int y = 0);
+    void new_sprite(char sprite, float x = 0, float y = 0);
+    void new_collectible(char sprite, float x = 0, float y = 0);
+    void new_end(char sprite, float x = 0, float y = 0);
+    void new_physics(char sprite, float x = 0, float y = 0);
     void new_floor(int left_x, int right_x, int y);
-    void new_player(char sprite, int x = 0, int y = 0);
+    void new_player(char sprite, float x = 0, float y = 0);
     bool tick(std::vector <int>);
 };
